+++ HISTORY.md
@@ -1,5 +1,6 @@
 ## 1.2.0 (2009-??-??)
 
+* If USR1 is sent and no child is found, shutdown.
 * Raise when a job class does not respond to `perform`.
 * Added `Resque.remove_queue` for deleting a queue